// src/types/PagedResponse.tsx
export interface PagedResponse<T> {
  content: T[];
  page: {
    size: number;
    number: number;
    totalElements: number;
    totalPages: number;
  };
}

export interface SpringPagedResponse<T> {
  content: T[];
  page: {
    size: number;
    number: number;
    totalElements: number;
    totalPages: number;
  };
}

export interface PlainPagedResponse<T> {
  content: T[];
  totalElements: number;
  totalPages: number;
}